cleanup + refactoring
