+++ tee-worker/ts-tests/integration-tests/common/utils/assertion.ts
@@ -251,10 +251,8 @@ export async function assertVc(context: IntegrationTestContext, subject: CorePri
     // check vc signature
     const signature = Buffer.from(hexToU8a(`0x${proof.proofValue}`));
     const message = Buffer.from(JSON.stringify(vcWithoutProof));
-
     const vcPubkeyBytes = context.api.createType('Option<Bytes>', lastRegisteredEnclave.vcPubkey).unwrap();
     const vcPubkey = Buffer.from(hexToU8a(vcPubkeyBytes.toHex()));
-
     const signatureStatus = await ed.verify(signature, message, vcPubkey);
 
     assert.isTrue(signatureStatus, 'Check Vc signature error: signature should be valid');
